refactor: moved decorator to its own folder
